Alt multi device

+++ Core_Program/remove_image.py
@@ -1,10 +1,10 @@
 import os
 
-def remove_mode():
+def remove_mode(current_IP):
     target_directory = './Main-Image-Captured/'
     image_list_directory = os.listdir(target_directory)
     for images in image_list_directory:
-        if images.endswith(".png"):
+        if images.startswith(current_IP) and images.endswith(".png"):
             target_image = target_directory + images
             os.remove(target_image)
